the one

+++ payCalculator.py
@@ -2,11 +2,12 @@ def calculatePay():
     # Implement your solution in between the two comment blocks
     print("calculating pay")
     # This first line is provided for you
-hrs = int(input("enter hours:"))
+hrs = float(input("enter hours:"))
 rate = float(input("enter rate:"))
 pay = (hrs*rate)
 
 print (pay)
+
